@@ -1,11 +1,13 @@
  class Order {
 
+
    constructor (
      public address:string,
      public number: number,
      public optionalAddress: string,
      public paymentOption: string,
-     public orderItems: OrderItem[]
+     public orderItems: OrderItem[],
+     public id?: string
    ){}
 
  }
